Update README link about making a PR
Under "Making contributions", the current link https://help.github.com/articles/creating-a-pull-request/ assumes that the branch is on the upstream repo. This could be confusing for students, who in the same section have just been told to make changes to a forked repo. Better to use https://help.github.com/articles/creating-a-pull-request-from-a-fork/ which will match what they have been doing.